+++ modules/common/src/main/java/net/oxcodsnet/beltborne_lanterns/common/server/BeltLanternServer.java
@@ -85,4 +85,3 @@ public final class BeltLanternServer {
         return null;
     }
 }
-